feat: move login-button to auth folder
